+++ modules/home-manager/default.nix
@@ -4,6 +4,7 @@
   git = ./git.nix;
   packages = ./packages.nix;
   development = ./development.nix;
+  codex = ./codex.nix;
 
   all = {
     imports = [
@@ -11,6 +12,7 @@
       ./git.nix
       ./packages.nix
       ./development.nix
+      ./codex.nix
     ];
   };
 }